\name{status.index}
\alias{status.index}
\title{Social Status Index}
\usage{
status.index(x)
}
\arguments{
  \item{x}{user record (logical vector) to calculate social
  status index from}
}
\description{
Calculates social status index, i.e. difference between
positive and negative votes count corrected by sample
length.
}

