\name{sociogram}
\alias{sociogram}
\title{.. content for \description{} (no empty lines) ..}
\usage{
sociogram(data, person.column, choice.columns, group.column, ...)
}
\arguments{
  \item{data}{}

  \item{person.column}{}

  \item{choice.columns}{}

  \item{group.column}{}

  \item{...}{}
}
\description{
.. content for \details{} ..
}

